requirements.txt
python-docx
requests
deep-translator
python-dotenv
